requirements.txt
django
djangorestframework
python-dotenv
django-crispy-forms
crispy-bootstrap5

#sortablejs
#webpack